% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardif.R
\name{cardif}
\alias{cardif}
\title{Compute absorption rate of diffuse radiation}
\usage{
cardif(leaf_albedo, ground_albedo, kdif)
}
\arguments{
\item{leaf_albedo}{The foliage albedo}

\item{ground_albedo}{The ground albedo}

\item{kdif}{diffuse radiation interception coefficient}
}
\value{
The absorption rate of diffuse radiation
}
\description{
This function returns absorption rate of diffuse radiation
}
\examples{
cardif(leaf_albedo = 0.22, ground_albedo = 0.18, kdif = 0.6)
}
