% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celestial_longitude.R
\name{celestial_longitude}
\alias{celestial_longitude}
\title{Compute celestial longitude of a date}
\usage{
celestial_longitude(date)
}
\arguments{
\item{date}{The date in format ("2019-05-30")}
}
\value{
The celestial longitude
}
\description{
This function returns celestial longitude from a date
}
\examples{
celestial_longitude("2019-05-30")
}
