% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_of_perigee.R
\name{day_of_perigee}
\alias{day_of_perigee}
\title{Compute day of perigee of a year}
\usage{
day_of_perigee(year)
}
\arguments{
\item{year}{The year in format (2019)}
}
\value{
The day of perigee
}
\description{
This function returns day of perigee of a year
}
\examples{
day_of_perigee(2019)
}
