% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equa_time.R
\name{equa_time}
\alias{equa_time}
\title{Compute equation of time of a date}
\usage{
equa_time(date)
}
\arguments{
\item{date}{The date in format ("2019-05-30")}
}
\value{
The equation of time
}
\description{
This function returns equation of  from a date
}
\examples{
equa_time("2019-05-30")
}
