% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redif.R
\name{redif}
\alias{redif}
\title{Compute reflexion rate of diffuse radiation}
\usage{
redif(leaf_albedo, ground_albedo, kdif)
}
\arguments{
\item{leaf_albedo}{The foliage albedo}

\item{ground_albedo}{The ground albedo}

\item{kdif}{diffuse radiation interception coefficient}
}
\value{
The reflexion rate of diffuse radiation
}
\description{
This function returns reflexion rate of diffuse radiation
}
\examples{
redif(leaf_albedo = 0.22, ground_albedo = 0.18, kdif = 0.6)
}
