% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb_to_albedo.R
\name{rgb_to_albedo}
\alias{rgb_to_albedo}
\title{Compute albedo from rgb soil color}
\usage{
rgb_to_albedo(r, g, b)
}
\arguments{
\item{r}{The r}

\item{g}{The b}

\item{b}{The g}
}
\value{
The albedo
}
\description{
This function returns albedo
}
\examples{
rgb_to_albedo(r=197,g=146,b=125)
}
