% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rurawls.R
\name{rurawls}
\alias{rurawls}
\title{rurawls
Calcul de la reserve utile selon la formule de Rawls et al. (1982)}
\usage{
rurawls(h, s, a, mo, pierosite, eg)
}
\arguments{
\item{h}{taille de l'horizon en m}

\item{s}{pourcentage de sables (0<x<100)}

\item{a}{pourcentage d'argiles (0<x<100)}

\item{mo}{pourcentage de matieres organiques  (0<x<100)}

\item{pierosite}{TRUE ou FALSE si la formule doit prendre en compte la pierosite (elements grossiers) ou non}

\item{eg}{proportion (0<x<1) d'elements grossiers}
}
\value{
reserve utile en mm
}
\description{
rurawls
Calcul de la reserve utile selon la formule de Rawls et al. (1982)
}
\examples{

rurawls(h=0.5,s=43,a=18,mo=1, pierosite=TRUE, eg = 0.4)

}
