% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vitirad.R
\name{vitirad}
\alias{vitirad}
\title{Compute solar radiation analysis of a vine location at a date}
\usage{
vitirad(lat, lon, date, tz, leaf_albedo, ground_albedo, height, thickness,
  interrow_width, porosity, row_aspect)
}
\arguments{
\item{lat}{The latitude in decimal degrees}

\item{lon}{The longitude in decimal degrees}

\item{date}{The date in format ("2019-05-30")}

\item{tz}{The timezone code see https://en.wikipedia.org/wiki/List_of_tz_database_time_zones}

\item{leaf_albedo}{The foliage albedo}

\item{ground_albedo}{The ground albedo}

\item{height}{The foliage height in meters}

\item{thickness}{The vineyard row thickness in meters}

\item{interrow_width}{The vineyard interrow width in meters}

\item{porosity}{The vine porosity (0<x<1)}

\item{row_aspect}{The row aspect (0 : NS, 90 : EW, -45 : NW-SE)}
}
\value{
The data frame of computed radiations
}
\description{
This function returns equation of from a date and the plot of the radiations
}
\examples{
vitirad(lat = 43.5, lon = 4.7,
        date = "2019-05-30", tz = "CET", leaf_albedo = 0.22,
        ground_albedo = 0.18, height = 1.4, thickness = 0.7,
        interrow_width = 2.4, porosity = 0.1, row_aspect = 0)
}
